requirements.txt
# Django核心
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# 数据库
mysqlclient==2.2.4
django-mysql==4.12.0

# 缓存
redis==5.0.1
django-redis==5.4.0

# 认证相关
PyJWT==2.8.0
cryptography==41.0.7

# 微信支付 (暂时注释，后续手动安装)
# WeChatPay==1.1.5
requests==2.31.0

# 图片处理
Pillow==10.1.0

# 环境变量管理
python-dotenv==1.0.0

# API文档
drf-spectacular==0.26.5

# 开发工具
django-debug-toolbar==4.2.0
django-extensions==3.2.3 